+ Environment Generator Bitmap Settings restructuring
