+++ Packages/com.pg649.worldgeneration/Runtime/Space Partitioning/Scenes/EnvironmentGeneratorBitmapSettings.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b8ccd8564c00849e08da08ef0161b3f2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
