+++ Packages/com.pg649.worldgeneration/Runtime/Space Partitioning/Scenes/TestScene.unity
@@ -174,6 +174,7 @@ MonoBehaviour:
   maxCorridorHeight: 3
   maxDistance: 32
   createTexture: 0
+  environmentGeneratorSettings: {fileID: 11400000, guid: b8ccd8564c00849e08da08ef0161b3f2, type: 2}
 --- !u!4 &59833268
 Transform:
   m_ObjectHideFlags: 0
