--- Packages/com.pg649.worldgeneration/Runtime/Terrain Generation/Scripts/BorderGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 349a9f7cc73f5df45a6987d0d4282191
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
