+++ Packages/com.pg649.worldgeneration/Runtime/Terrain Generation/Scripts/EnvironmentGeneratorBitmapSettings.asset
@@ -1,0 +1,19 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 56ebc2fc94bf14680a988fb668c4e2c4, type: 3}
+  m_Name: EnvironmentGeneratorBitmapSettings
+  m_EditorClassIdentifier: 
+  Depth: 100
+  AMPLITUDE: 0.8
+  FREQUENCY: 192
+  BlurringPasses: 1
+  BlurRadius: 10
