+++ Packages/com.pg649.worldgeneration/Runtime/Terrain Generation/Scripts/EnvironmentGeneratorBitmapSettings.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5fcced3ff789641fab8b67bbe71c5c66
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
