+++ Packages/com.pg649.worldgeneration/Runtime/Terrain Generation/Scripts/EnvironmentGeneratorBitmapSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 56ebc2fc94bf14680a988fb668c4e2c4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
