--- Packages/com.pg649.worldgeneration/Runtime/Terrain Generation/Scripts/ObstacleGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 42a7104512575ee42b3ec6d650773173
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
