--- Packages/com.pg649.worldgeneration/Runtime/Terrain Generation/Scripts/PerlinGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eef6336e230bb499a9ac8348aeed52ee
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
