--- Packages/com.pg649.worldgeneration/Runtime/Terrain Generation/Scripts/ZoneManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4e362f33933232a42bb52993f9cbf5df
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
