+++ Packages/com.pg649.worldgeneration/Runtime/Terrain Generation/Scripts/old_generator.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d61ef0a4f38f74d29af2d8d5b3ff51a3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
