+++ Packages/com.pg649.worldgeneration/Runtime/Terrain Generation/Scripts/old_generator/EnvironmentGenerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5e876df6e53795e47837310d1a57082f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
