+++ Packages/com.pg649.worldgeneration/Runtime/Terrain Generation/Scripts/old_generator/EnvironmentGeneratorSettings.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 42a3d3102f140cf41ac7a8b8be885e72
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
