+++ Packages/com.pg649.worldgeneration/Runtime/Terrain Generation/Scripts/old_generator/ObstacleGenerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 42a7104512575ee42b3ec6d650773173
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
